Move shared class into shared module
